+++ src/RequestConfig.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Shoxcie\BatchHttpClient;
 
 use Closure;
+use Symfony\Contracts\HttpClient\Exception\ExceptionInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
-use Throwable;
 
 final readonly class RequestConfig
 {
@@ -15,7 +15,7 @@ final readonly class RequestConfig
         public string         $url,
         /** @var array<string, mixed> */
         public array          $options = [],
-        /** @var array<string, mixed>|Closure(int, Throwable): array<string, mixed> */
+        /** @var array<string, mixed>|Closure(int, ExceptionInterface|InvalidResponseException): array<string, mixed> */
         public array|Closure  $retryOptions = [],
         public bool           $throwOnExhausted = true,
         public bool           $decodeJson = true,